+++ shoppinglist/database.py
@@ -26,3 +26,6 @@ def init_db():
     with current_app.open_resource("schema.sql") as f:
         db.executescript(f.read().decode("utf8"))
 
+
+def init_app(app):
+    app.teardown_appcontext(close_db)
